Fixed MoRE for OSX. Wrote build command in the iphone builder.
git-svn-id: svn://localhost/public/MoSync/trunk@1303 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/cpp/core/Core.cpp
@@ -389,7 +389,7 @@ public:
 					}
 					fprintf(file, " c=\"%i\" t=\"%f\" lt=\"%f\" ch=\"%i\" cht=\"%f\">\n", mCount,
 						mRunTime.toMilliSeconds(), (mRunTime - childrenTime).toMilliSeconds(),
-						mChildren.size(), childrenTime.toMilliSeconds());
+						(int)mChildren.size(), childrenTime.toMilliSeconds());
 					for(size_t i=0; i<mChildren.size(); i++) {
 						ProfNode* n = mChildren[i];
 						dLev++;
